requirements.txt
openai
PyYAML
numpy
matplotlib